+++ client/src/pages/MatchRegistration/Match.css
@@ -1,0 +1,218 @@
+@import url('https://fonts.googleapis.com/css2?family=Jersey+10&family=Pixelify+Sans:wght@400..700&display=swap');
+
+/* FORMULÁRIO */
+.match-form {
+    background-color: #1e1e2f;
+    color: #fff;
+    padding: 20px;
+    border-radius: 10px;
+    max-width: 600px;
+    margin: 40px auto;
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.4);
+    font-family: 'Pixelify Sans', sans-serif;
+    animation: subtleSlideIn 0.5s ease-out;
+}
+
+.match-form h2 {
+    text-align: center;
+    margin-bottom: 20px;
+    font-size: 24px;
+    font-weight: bold;
+}
+
+.input-group {
+    display: flex;
+    justify-content: space-between;
+    gap: 20px;
+    margin-bottom: 15px;
+}
+
+.input-group div {
+    flex: 1;
+}
+
+.input-group label {
+    display: block;
+    margin-bottom: 5px;
+    font-size: 14px;
+    text-transform: uppercase;
+    letter-spacing: 1px;
+}
+
+.input-group input {
+    width: 100%;
+    padding: 3px;
+    border: none;
+    border-radius: 5px;
+    background-color: #2e2e44;
+    color: #fff;
+    transition: background-color 0.3s ease, box-shadow 0.3s ease;
+}
+
+.input-group input:focus {
+    background-color: #3b3b5a;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.2);
+    outline: none;
+}
+
+/* BOTÕES */
+.button-group {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 15px;
+}
+
+.button-group button {
+    background-color: #5a4df2;
+    border: none;
+    border-radius: 5px;
+    color: #fff;
+    padding: 10px 20px;
+    cursor: pointer;
+    transition: transform 0.2s ease, background-color 0.3s ease, box-shadow 0.3s ease;
+}
+
+.button-group button:hover {
+    transform: scale(1.1);
+    background-color: #4e44d0;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
+}
+
+.button-group button:nth-child(2) {
+    background-color: #f25a5a;
+}
+
+.button-group button:nth-child(2):hover {
+    background-color: #d04646;
+    transform: scale(1.1);
+}
+
+/* TABELA */
+.matches-table {
+    margin: 20px auto;
+    max-width: 800px;
+    border-collapse: collapse;
+    font-family: 'Pixelify Sans', sans-serif;
+    color: #fff;
+    background-color: #1e1e2f;
+    border-radius: 10px;
+    overflow: hidden;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
+    animation: subtleSlideIn 0.6s ease-out;
+}
+
+.matches-table table {
+    width: 100%;
+    border-collapse: collapse;
+}
+
+.matches-table th, 
+.matches-table td {
+    padding: 12px 20px;
+    text-align: left;
+}
+
+.matches-table th {
+    background-color: #2e2e44;
+    font-size: 16px;
+    font-weight: bold;
+    text-transform: uppercase;
+    border-bottom: 2px solid #4e44d0;
+    opacity: 0;
+    animation: fadeIn 0.8s ease-out forwards 0.3s;
+}
+
+.matches-table td {
+    background-color: #2e2e44;
+    border-bottom: 1px solid #3a3a55;
+    transition: background-color 0.3s ease;
+}
+
+
+.matches-table th.actions {
+    text-align: right; 
+}
+
+
+.matches-table td.actions {
+    text-align: right; 
+    padding-right: 10px; 
+}
+
+.matches-table .actions {
+    display: flex;
+    justify-content: flex-end; 
+    gap: 10px;
+}
+
+.matches-table .actions button {
+    background-color: #5a4df2;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    padding: 4px 8px; 
+    font-size: 10px; 
+    line-height: 1.3; 
+    height: 24px;
+    width: auto;
+    cursor: pointer;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+}
+
+.matches-table .actions button:hover {
+    background-color: #4e44d0;
+    transform: scale(1.1); 
+}
+
+.matches-table .actions button.delete-button {
+    background-color: #f25a5a;
+}
+
+.matches-table .actions button.delete-button:hover {
+    background-color: #d04646;
+}
+
+/* ANIMAÇÕES */
+@keyframes subtleSlideIn {
+    0% {
+        transform: translateY(20px);
+        opacity: 0.8;
+    }
+    25% {
+        transform: translateY(15px);
+        opacity: 0.9;
+    }
+    50% {
+        transform: translateY(10px);
+        opacity: 0.95;
+    }
+    75% {
+        transform: translateY(5px);
+        opacity: 0.98;
+    }
+    100% {
+        transform: translateY(0);
+        opacity: 1;
+    }
+}
+
+@keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    50% {
+        opacity: 0.5;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@keyframes glowOnHover {
+    0% {
+        box-shadow: 0 0 5px rgba(255, 255, 255, 0.1);
+    }
+    100% {
+        box-shadow: 0 0 15px rgba(255, 255, 255, 0.3);
+    }
+}
